Adds context test scaffolding; renaming of files
- Renamed reducers and context to use hypen-style naming instead of
pascal case to avoid confusion

- Updated all references to reducer and context in all associated
components
- Adds basic unit test scaffolding to context
