+++ client/src/components/App.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import { darkTheme, lightTheme } from '../themes/theme';
-import { PackProvider } from '../contexts/packContext';
+import { PackProvider } from '../contexts/pack-context';
 
 import { ApolloClient, HttpLink } from 'apollo-boost';
 import { ApolloProvider } from 'react-apollo';
