+++ client/src/components/PackCarousel.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { usePackDispatch } from '../contexts/packContext';
+import { usePackDispatch } from '../contexts/pack-context';
 import { Mutation } from  'react-apollo';
 import { GET_FLASHCARDS_BY_PACK, CREATE_FLASHCARD} from '../queries';
