+++ client/src/components/PackHome.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import { usePackDispatch } from '../contexts/packContext';
+import { usePackDispatch } from '../contexts/pack-context';
 
 import LinkBackIcon from './icons/LinkBackIcon';
 import { SmallButton } from './elements/Button';
